fix(settings): fix deleting users with no drafts

+++ src/app/api/user/delete-account/route.ts
@@ -29,21 +29,24 @@ export async function POST() {
       throw new Error("Couldn't retrieve user data");
     }
 
+		// Delete drafts
     const draftMappings = userData.draftMappings;
-    draftMappings.forEach(async (mapping: { id: number; name: string }) => {
-      const query = db
-        .collection("drafts")
-        .where("draftId", "==", `${user.uid}-${mapping.id}`);
-      query.get().then((draftSnapshot) => {
-        draftSnapshot.docs.forEach((doc) => {
-          doc.ref.delete();
-        });
-      });
-    });
+		if (draftMappings) {
+			draftMappings.forEach(async (mapping: { id: number; name: string }) => {
+				const query = db
+					.collection("drafts")
+					.where("draftId", "==", `${user.uid}-${mapping.id}`);
+				query.get().then((draftSnapshot) => {
+					draftSnapshot.docs.forEach((doc) => {
+						doc.ref.delete();
+					});
+				});
+			});
+		}
 
     userDoc.delete();
 
-    // Destroy session cookie and log user out
+    // Destroy session cookie
     reqCookies.getAll().map((cookieKey) => {
       reqCookies.delete(cookieKey);
     });
